Update Node to 18 and package*.json to fix 'npm install'
The first problem: `docker compose build` does not tell us that `npm
install` actually had warnings and errors, which is very misleading. I
could not find a solution for this. To see the errors I had to access
the container's shell and run `rm -r node_modules; npm install` inside
it.

The second problem: the errors from npm install indicated that some
packages required npm >8.0.0, but it had npm around 6.0.0, and some
optional packages were not being installed. I just updated the
Dockerfile to use Node 18, since Node 14 and 16 are now deprecated, and
it got fixed.

The third problem: package.json package-lock.json were created from the
host, not from the container/docker enviroment. That led to some version
conflicts because the host's node and npm versions are different from
the ones in the container.

I could not find an "easy fix" for that, but for now I recommend
accessing the container's shell through
`docker compose run frontend /bin/bash`
and running npm commands in there everytime you want to deal with node
packages.

The new package*.json files in this commit were generated in that way.

Signed-off-by: Enzo Hideo <enzohideo@gmail.com>